Issue 29: Remaining IIR stuff (#31)
* Ringing

* Luma smear

* .gitignore

* Ditch breakpoints from git

* Get delay working

+++ ntsc-mtl-demo/Filters/NTSCEffect.swift
@@ -106,7 +106,7 @@ struct FBMNoiseSettings {
 struct RingingSettings {
     var frequency: Float
     var power: Float
-    var intensity: Float
+    var intensity: Float16
 }
 
 struct VHSSettings {
